% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaDecay.R
\name{alphaDecay}
\alias{alphaDecay}
\title{alphaDecay}
\usage{
alphaDecay(dataset.list, k = 5, alpha = 1, threshold = 1e-04)
}
\arguments{
\item{dataset.list}{A list (per batch) of CPCAs. See \link{cpca}}

\item{k}{The number of neighbors to use for the decay function [default = 5]}

\item{alpha}{The power of the decay function [default = 2]}

\item{threshold}{Threshold the minimum similarity between cells [default = 1e-4]}
}
\value{
List (per-batch) of alpha decay kernels.
}
\description{
Function to compute the alpha decay kernels
}
