% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSNN.R
\name{computeSNN}
\alias{computeSNN}
\title{computeSNN}
\usage{
computeSNN(dataset.list, k = "rice", prune = 1/15, add = 20, rare = TRUE)
}
\arguments{
\item{dataset.list}{A list (per batch) of CPCAs. See \link{cpca}}

\item{k}{The number of neighbors to use for building the SNN graph [default = "rice" (= (2 * n^(1/3)))]}

\item{prune}{The minimum number of normalized neighbors between cells to retain the eddge [default = 1/15]}

\item{add}{The number of neighbors to add in addition to k in neighbourhood search [default = 20]}

\item{rare}{A boolean (TRUE or FALSE) indicating if rare cell index should be calculated from the distance to the Kth neighbour [default = TRUE]}
}
\value{
A list containing a list (per batch) of SNN graphs and a list (per-batch) of rare cell indices
}
\description{
Function to compute and prune an SNN graph
}
