% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPCA.R
\name{cpca}
\alias{cpca}
\title{CPCA}
\usage{
cpca(
  dataset.list,
  weight_by_var = TRUE,
  pca.type = "cpca",
  nfeat = 1000,
  selection.method = "deviance",
  threshold = 0.8,
  kc = 20,
  ki = 20,
  do.center = FALSE,
  feat.type = "inclusive",
  oversampling = 10,
  ncpu = 1,
  iter.max = 100,
  tol = 1e-05,
  eps = .Machine$double.eps^(4/5),
  verbose = TRUE
)
}
\arguments{
\item{dataset.list}{A list (per batch) of raw counts}

\item{weight_by_var}{Boolean (TRUE or FALSE) indicating whether or not to weight embeddings by the variance of each PC [default = TRUE]}

\item{pca.type}{The method (either 'cpca' or 'irlba') for computing PCs [default = "cpca"]}

\item{nfeat}{The number of features to select [default = 1000]}

\item{selection.method}{The method to use selecting highly-variable features [default = "deviance"]}

\item{threshold}{The minimum amount of variance to be explained by common PCs before adding individual PCs [default = 0.8]}

\item{kc}{The number of common PCs to calculate [default = 20]}

\item{ki}{The number of invididual PCs to calcuate [default = 20]}

\item{do.center}{Boolean indicating whether or not to center expression values prior to decomposition [default = FALSE]}

\item{feat.type}{The method (either 'inclusive' or 'exclusive') for selecting variable feature [default = "inclusive"]}

\item{oversampling}{The number of PCs to calculate in excess the selected PCs [default = 10]}

\item{ncpu}{The number of cpus to use for matrix multiplications [default = 1]}

\item{iter.max}{The maximum number of iterations to run svd for [default = 100].}

\item{tol}{Tolerance (minimum increase in fit) for early stopping svd decomposition [default = 1e-5]}

\item{eps}{Epsilon (minimum values of distances) for early stopping svd decomposition [default = .Machine$double.eps ^ (4 / 5)]}

\item{verbose}{Boolean (TRUE or FALSE) indicating to print messages [default = TRUE]}
}
\value{
List containing a list (per-batch) of CPCAs and a list (per-batch) of normalized counts.
}
\description{
Function to perform consensus PCA
}
