% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOINTLYsolve.R
\name{JOINTLYsolve}
\alias{JOINTLYsolve}
\title{JOINTsolve}
\usage{
JOINTLYsolve(
  kernel.list,
  snn.list,
  rare.list,
  cpca.result,
  init = "clustering",
  norm.scale = TRUE,
  norm.minmax = FALSE,
  norm.center = FALSE,
  k = 15,
  m = 2,
  iter.max = 200,
  alpha = 1,
  mu = 20,
  lambda = 1,
  beta = 5,
  progressbar = TRUE,
  share.objects = TRUE,
  ncpu = 1,
  save_all = FALSE,
  bpparam = SerialParam()
)
}
\arguments{
\item{kernel.list}{A list (per-batch) of kernels. See \link{alphaDecay}}

\item{snn.list}{A list (per-batch) of SNN graphs. See \link{computeSNN}}

\item{rare.list}{A list (per-batch) of rare cell indices. See \link{computeSNN}}

\item{cpca.result}{The results from CPCA analysis. See \link{cpca}}

\item{init}{The name of the method to use for initialization ("random" or "clustering"). [default = "clustering"]}

\item{norm.scale}{Boolean (TRUE or FALSE) indicating whether or not to scale the normalized counts [default = TRUE]}

\item{norm.minmax}{Boolean (TRUE or FALSE) indicating whether or not to MinMax scale the normalized counts [default = FALSE]}

\item{norm.center}{Boolean (TRUE or FALSE) indicating whether or not to center the normalized counts [default = FALSE]}

\item{k}{The number of factors to calculate [default = 15]}

\item{m}{The fuzzy parameter of for cluster initialization [default = 2]}

\item{iter.max}{The maximum number of iterations to perform [default = 200]}

\item{alpha}{Alpha parameter of the loss function [default = 1]}

\item{mu}{Mu parameter of the loss function [default = 20]}

\item{lambda}{Lambda parameter of the loss function [default = 1]}

\item{beta}{Beta parameter of the loss function [default = 5]}

\item{progressbar}{A logical indicating whether or not to print a progress bar [default = TRUE]}

\item{share.objects}{Boolean (TRUE or FALSE) indicating to use shared object to reduce memory requirements for parallel processing [default = TRUE]}

\item{ncpu}{The number of cpus to use for matrix multiplication [default = 1]}

\item{save_all}{Boolean (TRUE or FALSE) indicating if all H, F and W matrices should be saved and outputted [default = FALSE]}

\item{bpparam}{*Param to use for parallel processing [default = SerialParam()]}
}
\value{
A list containing normalized H matrices to use for clustering, as well a list (per-batch) of H matrices, a list (per-batch) of F matrices and a list (per-batch) of W matrices.
}
\description{
Function to solve the NMF problem in JOINTLY
}
