% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  dataset.list,
  k.decay = 10,
  k.rare = 20,
  k.snn = 100,
  alpha = 2,
  threshold = 1e-04,
  prune = 1/15,
  rare = TRUE
)
}
\arguments{
\item{dataset.list}{A list (per batch) of CPCAs. See \link{cpca}}

\item{k.decay}{The number of neighbors to use for the decay function [default = 10]}

\item{k.rare}{The index of the nearest neighbor to use for rare scoring [default = 20]}

\item{k.snn}{The number of neighbors to use for the SNN calculations [default = 100]}

\item{alpha}{The power of the decay function [default = 2]}

\item{threshold}{Threshold the minimum similarity between cells [default = 1e-4]}

\item{prune}{The minimum number of normalized neighbors between cells to retain the eddge [default = 1/15]}

\item{rare}{A boolean (TRUE or FALSE) indicating if rare cell index should be calculated from the distance to the Kth neighbour [default = TRUE]}
}
\value{
List (per-batch) of alpha decay kernels.
}
\description{
Function to compute the alpha decay kernels, rareity scores and SNN graphs
}
