% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess}
\usage{
preprocess(data, batch.var = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A single (or list of) Seurat, SingleCellExperiment, or Matrix-like object(s) containing raw counts.}

\item{batch.var}{A variable to split Seurat or SingleCellExperiment objects into a list [default = NULL]}

\item{verbose}{Boolean (TRUE or FALSE) indicating to print messages [default = TRUE]}
}
\value{
A list (per-batch) of raw counts
}
\description{
Function to preprocess data from Seurat, SCE or Matrix-like objects
}
