% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaDecay.R
\name{alphaDecay}
\alias{alphaDecay}
\title{alphaDecay}
\usage{
alphaDecay(dataset.list, k = 5, alpha = 2, theta = 1e-04)
}
\arguments{
\item{dataset.list}{A list (per batch) of CPCAs. See \link{cpca}}

\item{k}{The number of neighbors to use for the decay function [default = 5]}

\item{alpha}{The power of the decay function [default = 2]}

\item{theta}{A number modifying the bandwidth of the decay function [default = 1e-4]}
}
\value{
List (per-batch) of alpha decay kernels.
}
\description{
Function to compute the alpha decay kernels
}
