% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jointly}
\alias{jointly}
\title{jointly}
\usage{
jointly(
  data,
  batch.var = NULL,
  factors = 20,
  nfeat = 1000,
  selection.method = "deviance",
  decay.k = 5,
  decay.alpha = 2,
  cpca.threshold = 0.8,
  cpca.kc = 20,
  cpca.ki = 20,
  alpha.loss = 100,
  mu.loss = 1,
  lambda.loss = 100,
  beta.loss = 1,
  snn.k = "rice",
  snn.add = 20,
  snn.rare = TRUE,
  ncpu = 1,
  iter.max = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A single (or list of) Seurat, SingleCellExperiment, or Matrix-like object(s) containing raw counts.}

\item{batch.var}{A variable to split Seurat or SingleCellExperiment objects into a list [default = NULL]}

\item{factors}{The number of factors to calculate [default = 20]}

\item{nfeat}{The number of features to select [default = 1000]}

\item{selection.method}{The method to use selecting highly-variable features [default = "deviance"]}

\item{decay.k}{The number of neighbors to use for the decay function [default = 5]}

\item{decay.alpha}{The power of the decay function [default = 2]}

\item{cpca.threshold}{The minimum amount of variance to be explained by common PCs before adding individual PCs [default = 0.8]}

\item{cpca.kc}{The number of common PCs to calculate [default = 20]}

\item{cpca.ki}{The number of invididual PCs to calcuate [default = 20]}

\item{alpha.loss}{Alpha parameter of the loss function [default = 100]}

\item{mu.loss}{Mu parameter of the loss function [default = 1]}

\item{lambda.loss}{Lambda parameter of the loss function [default = 100]}

\item{beta.loss}{Beta parameter of the loss function [default = 1]}

\item{snn.k}{The number of neighbors to use for building the SNN graph [default = "rice" (= (2 * n^(1/3)))]}

\item{snn.add}{The number of neighbors to add in addition to k to use for building the SNN graph [default = 20]}

\item{snn.rare}{A boolean (TRUE or FALSE) indicating if rare cell index should be calculated from the distance to the Kth neighbour during SNN graph construction [default = TRUE]}

\item{ncpu}{The number of cpus to use for matrix multiplication [default = 1]}

\item{iter.max}{The maximum number of iterations to perform [default = 100]}

\item{verbose}{Boolean (TRUE or FALSE) determining verbosity [default = TRUE]}

\item{...}{Additional parameters to pass to functions within JOINTLY}
}
\value{
A list containing a list (per-batch) of H matrices, a list (per-batch) of F matrices and a list (per-batch) of W matrices.
}
\description{
Function to cluster single-cell and single-nucleus RNA-seq data using JOINTLY.
}
