% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOINTLYsolve.R
\name{JOINTLYsolve}
\alias{JOINTLYsolve}
\title{JOINTsolve}
\usage{
JOINTLYsolve(
  kernel.list,
  snn.list,
  rare.list,
  norm.list,
  k = 20,
  iter.max = 100,
  alpha = 100,
  mu = 1,
  lambda = 100,
  beta = 1,
  ncpu = 1
)
}
\arguments{
\item{kernel.list}{A list (per-batch) of kernels. See \link{alphaDecay}}

\item{snn.list}{A list (per-batch) of SNN graphs. See \link{computeSNN}}

\item{rare.list}{A list (per-batch) of rare cell indices. See \link{computeSNN}}

\item{norm.list}{A list (per-batch) of normalized counts for highly variable genes. See \link{cpca}}

\item{k}{The number of factors to calculate [default = 20]}

\item{iter.max}{The maximum number of iterations to perform [default = 100]}

\item{alpha}{Alpha parameter of the loss function [default = 100]}

\item{mu}{Mu parameter of the loss function [default = 1]}

\item{lambda}{Lambda parameter of the loss function [default = 100]}

\item{beta}{Beta parameter of the loss function [default = 1]}

\item{ncpu}{The number of cpus to use for matrix multiplication [default = 1]}
}
\value{
A list containing a list (per-batch) of H matrices, a list (per-batch) of F matrices and a list (per-batch) of W matrices.
}
\description{
Function to solve NMF problems
}
